requirements.txt
# meaning of life - not found
discord.py==1.7.3
PyNaCl==1.5.0
python-dotenv==0.20.0
youtube-dl==2021.12.17